dt-bindings: mediatek: pwrap: Convert to DT schema
Convert PMIC wrapper DT bindings document to DT schema format.
The original author of the PMIC wrapper driver is added as
a maintainer.

Looking at the driver. the resets and reset-names are actually
required in some SoCs and totally unused in others. Instead of
making those properties optional like in the old document, add
a conditional to make them required or remove them based on the
compatible string.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>